Update markdown docs for GitHub Pages

+++ library/graph/minimumsteinertree.hpp
@@ -2,6 +2,7 @@
 #include "graphtemplate"
 #include <bits/stdc++.h>
 using namespace std;
+// 最小シュタイナー木を求める O(n*3^k + m*2^t log n) ただし、負閉路が存在しないこと
 template <class T, bool directed = false, bool weighted = true>
 graph<T, false, true> minimumsteinertree(graph<T, directed, weighted> &g, vector<int> &v) {
     vector<vector<T>> dp(1<<v.size(), vector<T>(g.size(), numeric_limits<T>::max()));
@@ -16,15 +17,15 @@ graph<T, false, true> minimumsteinertree(graph<T, directed, weighted> &g, vector
             id[_e.to][_e.from] = _e.id;
         }
     }
-    for (int i=0; i<g.size(); i++) {
+    for (int i=0; i<(int)g.size(); i++) {
         d[i][i] = 0;
     }
-    for (int i=0; i<v.size(); i++) {
+    for (int i=0; i<(int)v.size(); i++) {
         dp[1<<i][v[i]] = 0;
     }
     for (int i=0; i<(1<<v.size()); i++) {
         for (int j=i; 0<j; j=(j-1)&i) {
-            for (int k=0; k<g.size(); k++) {
+            for (int k=0; k<(int)g.size(); k++) {
                 if (dp[j][k] == numeric_limits<T>::max() || dp[i^j][k] == numeric_limits<T>::max()) continue;
                 if (dp[j][k] + dp[i^j][k] < dp[i][k]) {
                     dp[i][k] = dp[j][k] + dp[i^j][k];
@@ -33,7 +34,7 @@ graph<T, false, true> minimumsteinertree(graph<T, directed, weighted> &g, vector
             }
         }
         priority_queue<pair<T, int>, vector<pair<T, int>>, greater<pair<T, int>>> q;
-        for (int j=0; j<g.size(); j++) {
+        for (int j=0; j<(int)g.size(); j++) {
             if (dp[i][j] != numeric_limits<T>::max()) q.push({dp[i][j], j});
         }
         while (!q.empty()) {
@@ -50,7 +51,7 @@ graph<T, false, true> minimumsteinertree(graph<T, directed, weighted> &g, vector
     }
     int c = -1;
     T ans = numeric_limits<T>::max();
-    for (int i=0; i<g.size(); i++) {
+    for (int i=0; i<(int)g.size(); i++) {
         if (dp.back()[i] < ans) {
             ans = dp.back()[i];
             c = i;
